requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
